code refactoring to include all cards
